Remove any reference to the metadatafield 'subject'
The field 'subject' is a special case of opencast. The external API
however only returns the field as 'subjects' and it also accepts the
field as 'subjects'. I would not consider the special case of 'subject'
in singular in this plugin, since it is also not documented in the
external API, but exists solely in code.