Update list of default roles which are added to a series
This patch removes ROLE_GROUP_MOODLE_COURSE_[COURSEID] as role which is added to a series by default as it is not relevant for default installations.
On the other hand, it adds ROLE_GROUP_MH_DEFAULT_ORG_EXTERNAL_APPLICATIONS as role which is added to a series by default as this role is needed to let Moodle fetch series information.

This change is only applied to new installations.
Existing installations which are mostly configured by experienced admins in custom ways are not changed.