Overview page - Remove unused code to download the recordings table
There was code to download the recordings table in index.php, but the table was marked as not downloadable and thus the code was not used.
Even more, if you downloaded the table data forcibly, you just got broken data.
As there is no real use to download the recordings table, the download code fragments are removed.