Add LTI episode module feature
If enabled, teachers can add Opencast LTI episode modules to the course. These LTI episode modules will be pointing to an Opencast episode.

Note: This patch is built based on and in addition to the existing 'Add LTI series module' feature. To distinguish both features clearly, some existing code of the existing feature is modified.