Add LTI series module: Drop courseid unique key
With this patch, we drop the 'courseid' unique key and the 'fk_course' foreign-unique key from the table block_opencast_ltimodule.
Afterwards, we recreate the 'fk_course' foreign-unique key.
This is done as a foreign-unique key produces a unique key automatically and we do not need two identical keys.
This is especially done in this way (dropping both keys, then recreating the second one from scratch) as it can happen
that Moodle drops both keys already when you just want to drop the first one as both are technically the same.