Relabel the feature to delete the video before the transfer to Opencast
This commit changes a string for a label and heading and button which said "Delete video permanently".
As this feature is used to delete videos before the transfer to Opencast, the substring "permanently" is not really accurate and will be improved with this change to "Delete video before transfer to Opencast".