Relabel the feature to delete the video before the transfer to Opencast
This commit improves the strings of the feature which allows the teacher to delete a video before the transfer to Opencast.
This feature was labeled "Delete video permanently", but the substring "permanently" is not really accurate as the video is still in the draft status and not yet available for anyone at all.

In addition to that, this patch separates all strings of this feature from the strings of the permanent deletion feature.

In addition to that, this patch also adds some explanation to the deletedraft.php page to better explain to the teacher what will happen as soon as he confirms the deletion.