Add manual import videos feature - Solves #13, solves #21
This patch adds the possibility for teachers to import videos from other courses.
On the Opencast videos overview page, there is a new button which leads you to a new import wizard.
The wizard asks for the source course which hold the videos to be imported and for the videos to be imported from this course.

After confirming the import wizard, the teacher is redirected to the overview page again.
In the background, ad hoc tasks for duplicating the events are created, just like the existing Moodle core course import wizard does.

This patch is fully working, but it should be considered as incomplete in terms of
a) Handling of existing Opencast LTI modules. This is prepared in the wizard, but needs some logic.
b) Handling of really long source course lists. A filtering mechanism might make sense.