Move duplicateworkflow setting and create kill switch for Moodle course import wizard - Solves #159
This patch moves the duplicateworkflow setting to the newly created import videos section within the admin setting.
Additionally, it introduces a kill switch for importing videos with the Moodle core course import wizard.
The admin will then have the possibility to choose which of the two import solutions he wants to provide and support.

As an additional goodie, this patch introduced a special admin setting class which valides if the given workflow exists in Opencast.