Enhance get_stored_seriesid() function to create series on-the-fly
This patch adds the $createifempty parameter to the get_stored_seriesid() function in the apibridge class.
The parameter is set to false by default to keep the behaviour of existing calls of this function.

As soon you set the parameter to true when calling the function, the function creates a series on-the-fly if no series is mapped yet to the course.
With this enhancement, the caller does not need to create the series himself if he needs to ensure that a series exists.

This patch changes the obvious locations where this behaviour is needed in addlti.php and recordvideo.php.