Add possibility to cleanup episode modules within video import wizard
This patch enhances the video import wizard with an additional option which lets teachers cleanup existing episode modules.
This is especially helpful to fix episode modules which have been imported with the Moodle core course import wizard and which are pointing to an episode in the wrong course afterwards.

In contrast to the feature to cleanup series modules, the episode cleanup cannot be done directly after the video import wizard is finished. Instead, there is a scheduled task which waits for Opencast to do its job and will then cleanup the modules in the background. This scheduled task is enabled by default.