Use Moodle core course selector within import video wizard
This patch rebuilds step 1 or the import video wizard to use the course selector component which is used within the Moodle core course import wizard.

With this change, we get rid of the shortcomings of the existing implementation of the course selector. On the other hand, we are not able anymore to only offer courses which have videos to be imported.