Prevent that the plugin settings page breaks if Opencast API is not available
After the setting block_opencast | duplicateworkflow was moved to the plugin's "Additional features" settings page, the plugin was not installable anymore in new installations.
This was because the block_opencast | duplicateworkflow tried to retrieve the list of available workflows from Opencast which only works if tool_opencast is already configured.

This patch changes this is a way that admin_setting_configselect_opencastworkflow settings only load their choices if Moodle is already set up and in a way that they fail badly if the Opencast workflows can't be retrieved.

Additionally, the check if the Opencast API is configured was added as criteria to the check which happens before the widgets for the LTI module and import video features are shown.