Split additional features settings page
This patch splits the existing additional features settings into multiple pages: Additional features, LTI modules features and Import videos features.
This was necessary as the hide_if functionality for admin settings is only able to handle a limited amount if constraints per page. As soon as there are too many constraints, this triggers a debug warning and this lets Behat tests fail as they rely on the fact the no single debug message should be thrown.