Improve Opencast overview page UX when no videos have been uploaded yet - Solves #197.
This patch modifies the 'nothingtodisplay' string which is shown on the Opencast overview page when no videos have been uploaded yet with the goal to improve understandability.
It does not touch the 'novideosavailable' string which is shown on the block itself as this string should remain as short as possible.