When creating Opencast modules, respect the $CFG->completiondefault setting for completion tracking - Fixes #189.
Before this patch, even if the default in $CFG->completiondefault was set to "Use activity default", the created Opencast activities always used "Do not indicate activity completion".

This patch changes this shortcoming in a way that the Opencast module creation process will now respect and use the $CFG->completiondefault setting for completion tracking.