Pre-fill activity descriptions with existing episode descriptions - Solves #196.
When uploading a video on addvideo.php (and if configured properly by the Moodle admin), teachers can not only an episode title but also a episode description within the episode metadata.
Additionally, when providing the episode as (LTI or mod_opencast) activity to the students, the teacher can again add a description.

This patch adds the behaviour that an existing description from the episode metadata will be pre-filled into the activity description field.